refactor(client): move welcome to components
